fix: make shared (SSR) shelf rows size to viewer width
The shared /s/ page used ShelfGridStatic with a fixed items-per-row count
(no measurement), so its rows still wrapped onto a second visual line under
one ledge — e.g. at a ~1064px container only 6 fit but 7 were rendered, and
on mobile a 7-card row wrapped badly.

Move the row grouping into ShelfRowsResponsive, a thin client component that
sizes rows to the container via the shared useShelfRows hook. ItemCardStatic
still renders the item markup, so the server HTML keeps all <article>
nodes + data-item-id (crawler/AI readability and SharedShelfInteractive's
click delegation are unchanged); only the row split becomes responsive after
hydration. ShelfGridStatic stays a server component and keeps the
server-rendered empty state.

Verified: shared page now shows 6/row x2 (1280px) and 3/row x4 (375px) with
no wrapping, and the raw SSR HTML still contains all 12 articles + titles.
